Split app_segments into multiple source files (#1651)
Split app_segments into multiple .c files since it was getting pretty
massive (and I still need to add a bunch of stuff). Also added voltage
and temp stats.
